:art: Data Access 분리
